services/requirements.txt
fastapi
uvicorn
pytest
requests
pydantic
pyyaml
gradio
